requirements.txt
pip>=25.1
aiohttp==3.9.5
aiogram==2.25.1
python-dotenv
